reorganising views into subpages
