+++ website/__init__.py
@@ -45,7 +45,9 @@ def create_app():
     app.config['SECRET_KEY'] = str(uuid.uuid1())
 
     from .views import views
+    from .registrations import registrations
 
     app.register_blueprint(views, url_prefix='/')
+    app.register_blueprint(registrations, url_prefix='/')
 
     return app
